+++ sos21-database/src/query/list_projects.rs
@@ -1,6 +1,6 @@
 use crate::model::{
     project::{Project, ProjectAttributes, ProjectCategory, ProjectWithOwners},
-    user::{User, UserRole},
+    user::{User, UserCategory, UserRole},
 };
 
 use anyhow::{Context, Result};
@@ -35,6 +35,7 @@ SELECT
         owners.affiliation AS "owner_affiliation!",
         owners.email AS "owner_email!",
         owners.role AS "owner_role!: UserRole",
+        owners.category AS "owner_category!: UserCategory",
         subowners.created_at AS "subowner_created_at!",
         subowners.first_name AS "subowner_first_name!",
         subowners.kana_first_name AS "subowner_kana_first_name!",
@@ -43,7 +44,8 @@ SELECT
         subowners.phone_number AS "subowner_phone_number!",
         subowners.affiliation AS "subowner_affiliation!",
         subowners.email AS "subowner_email!",
-        subowners.role AS "subowner_role!: UserRole"
+        subowners.role AS "subowner_role!: UserRole",
+        subowners.category AS "subowner_category!: UserCategory"
 FROM projects
 INNER JOIN users AS owners ON (projects.owner_id = owners.id)
 INNER JOIN users AS subowners ON (projects.subowner_id = subowners.id)
@@ -78,6 +80,7 @@ INNER JOIN users AS subowners ON (projects.subowner_id = subowners.id)
             affiliation: row.owner_affiliation,
             email: row.owner_email,
             role: row.owner_role,
+            category: row.owner_category,
         };
         let subowner = User {
             id: row.subowner_id,
@@ -90,6 +93,7 @@ INNER JOIN users AS subowners ON (projects.subowner_id = subowners.id)
             affiliation: row.subowner_affiliation,
             email: row.subowner_email,
             role: row.subowner_role,
+            category: row.subowner_category,
         };
 
         Ok(ProjectWithOwners {